+++ services/src/main/java/org/keycloak/broker/saml/federation/SAMLFederationProvider.java
@@ -108,7 +108,7 @@ public class SAMLFederationProvider extends AbstractIdPFederationProvider <SAMLF
 	private static final String CATEGORY_CLIENTS= "Clients";
 	private static final String CATEGORY_IDPS= "Identity Providers";
 	private static final int MAX_LOGO_LENGTH = 4000;
-	private static final int DEFAULT_BATCH_SIZE = 3000;
+	private static final int DEFAULT_BATCH_SIZE = 1000;
 	private List<ProtocolMapperModel> defaultSAMLMappers;
 
 	public SAMLFederationProvider(KeycloakSession session, SAMLFederationModel model, String realmId) {